jsonrpsee

+++ client/src/http_sender.rs
@@ -74,7 +74,7 @@ impl HttpSender {
 
 #[derive(Deserialize, Debug)]
 pub(crate) struct RpcErrorObject {
-    pub code: i64,
+    pub code: i32,
     pub message: String,
 }
